pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clouddeploy"
version = "0.1.0"
description = "A self-hostable CLI deployment platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "docker>=7",
    "pyyaml>=6",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.scripts]
clouddeploy = "clouddeploy.main:app"

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-mock>=3", "ruff>=0.4"]

[tool.ruff]
line-length = 88
target-version = "py311"
